pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "TextArena"
version = "0.6.0"
description = "[WIP] A Collection of Competitive Text-Based Games for Language Model Evaluation and Reinforcement Learning"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    { name="Leon Guertler", email="Guertlerlo@cfar.a-star.edu.sg" },
]
dependencies = [
    "requests",
    "bitsandbytes==0.41.1",
    "openai",
    "rich",
    "nltk",
    "chess",
    "networkx",
    "python-dotenv",
    "opencv-python",
    "websockets",
    "mss",
    "fantasynames>=0.0.1",
    "transformers>=4.51.3",
    "torch>=2.6.0",
    "accelerate>=1.6.0",
    "anthropic>=0.49.0",
    "ipython>=8.35.0",
    "honcho-ai>=0.0.19",
]

[project.urls]
"Homepage" = "https://github.com/LeonGuertler/TextArena"
